+++ intellij-feature-extractor/src/test/java/com/intellij/featureExtractor/MockIdePlugin.kt
@@ -3,15 +3,18 @@ package com.intellij.featureExtractor
 import com.jetbrains.plugin.structure.base.plugin.PluginIcon
 import com.jetbrains.plugin.structure.base.plugin.ThirdPartyDependency
 import com.jetbrains.plugin.structure.intellij.plugin.Classpath
+import com.jetbrains.plugin.structure.intellij.plugin.DependsPluginDependency
 import com.jetbrains.plugin.structure.intellij.plugin.IdePlugin
 import com.jetbrains.plugin.structure.intellij.plugin.IdePluginContentDescriptor
 import com.jetbrains.plugin.structure.intellij.plugin.IdeTheme
 import com.jetbrains.plugin.structure.intellij.plugin.KotlinPluginMode
 import com.jetbrains.plugin.structure.intellij.plugin.Module
+import com.jetbrains.plugin.structure.intellij.plugin.ContentModuleDependency
 import com.jetbrains.plugin.structure.intellij.plugin.ModuleDescriptor
 import com.jetbrains.plugin.structure.intellij.plugin.MutableIdePluginContentDescriptor
 import com.jetbrains.plugin.structure.intellij.plugin.OptionalPluginDescriptor
 import com.jetbrains.plugin.structure.intellij.plugin.PluginDependency
+import com.jetbrains.plugin.structure.intellij.plugin.PluginMainModuleDependency
 import com.jetbrains.plugin.structure.intellij.plugin.ProductDescriptor
 import com.jetbrains.plugin.structure.intellij.version.IdeVersion
 import org.jdom2.Document
@@ -26,6 +29,9 @@ data class MockIdePlugin(
   override val appContainerDescriptor: IdePluginContentDescriptor = MutableIdePluginContentDescriptor()
   override val projectContainerDescriptor: IdePluginContentDescriptor = MutableIdePluginContentDescriptor()
   override val moduleContainerDescriptor: IdePluginContentDescriptor = MutableIdePluginContentDescriptor()
+  override val dependsList: List<DependsPluginDependency> = emptyList()
+  override val pluginMainModuleDependencies: List<PluginMainModuleDependency> = emptyList()
+  override val contentModuleDependencies: List<ContentModuleDependency> = emptyList()
   override val pluginName: String? = null
   override val description: String? = null
   override val url: String? = null
@@ -36,13 +42,15 @@ data class MockIdePlugin(
   override val icons: List<PluginIcon> = emptyList()
   override val productDescriptor: ProductDescriptor? = null
   override val dependencies: List<PluginDependency> = emptyList()
-  override val incompatibleModules: List<String> = emptyList()
+  override val incompatibleWith: List<String> = emptyList()
+  override val pluginAliases: Set<String> = emptySet()
   override val underlyingDocument: Document = Document(Element("idea-plugin"))
   override val optionalDescriptors: List<OptionalPluginDescriptor> = emptyList()
   override val modulesDescriptors: List<ModuleDescriptor> = emptyList()
   override val sinceBuild: IdeVersion = IdeVersion.createIdeVersion("IU-163.1")
   override val untilBuild: IdeVersion? = null
-  override val definedModules: Set<String> = emptySet()
+  @Deprecated("use either pluginAliases or contentModules")
+  override val definedModules: Set<String> = pluginAliases
   override val contentModules: List<Module> = emptyList<Module>()
   override val originalFile: Path? = null
   override val useIdeClassLoader = false