+++ intellij-plugin-structure/structure-ide/src/main/java/com/jetbrains/plugin/structure/ide/layout/ModuleFactory.kt
@@ -38,7 +38,6 @@ internal class ModuleFactory(private val moduleLoader: LayoutComponentLoader, pr
         IdeModule
           .clone(moduleLoadingResult.plugin, moduleName, classpath = getClasspath(moduleName, idePath))
           .apply {
-            definedModules += moduleName
             moduleDependencies += moduleDescriptor.dependencies
             resources += moduleDescriptor.resources
           }