+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/beans/PluginAliasBean.java
@@ -6,6 +6,6 @@ package com.jetbrains.plugin.structure.intellij.beans;
 
 import javax.xml.bind.annotation.XmlAttribute;
 
-public class ListItemBean {
+public class PluginAliasBean {
   @XmlAttribute(name = "value") String value;
 }